requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
asyncpg==0.28.0
pydantic==2.4.2
scipy==1.11.3
aiohttp==3.8.6
python-dotenv==1.0.0
